doc-service/requirements.txt
fastapi
uvicorn
python-docx
psycopg2-binary
